Created centralized Shape class
Created a Shape class, which currently can be initiated as a circle
or as a rectangle.  Removed the Circle and the Rectangle classes,
as they are now entirely obsolete.  Will better implement the
Initialization functions, possibly as class friend functions which
return the reference to the initialized object.  In either case,
also create a polygon initialization function.