Added shapes vector, created minimap
Added vector of massless shapes, which, for now, are created with
a right click.
Updated Shape's rectangle creation to take angle into account
Created a minimap, which is centered on the player, but remains a
constant angle.
Updated Player's footbox to be a translucent rectangle, in order to
better behave with a minimap.
Updated Render function on each state to set the view as desired
with each cycle, so that rendering doesn't go crazy.