Added Credits state, added background music
Added a credits state, which just gives credit to myself and the
art community for now.  Will be more specific later.
Added background music, pulled from Jamendo.  The file is somewhat
large, but as the format is rather specific, and the file will not
change often, I feel that committing it is warranted.  May be moved
to a separate downloadable package in the future, if files get too
large.