Added GetHiddenState, fixed player movement, updated PauseMenuState
Added GetHiddenState to GameCore.  This will retrieve the second
state from the top of the stack.  Here, this is used for rendering
the main game under the pause menu.
Fixed the player's movement states.  Instead of an all
encompassing enumeration, there is a movement enumeration and a
jumping bool.  This is used so that a player's keystrokes will not
be forgotten upon jumping.  This still is slightly buggy with
the pause menu and with switching windows, as it is still event-
based.
Updated PauseMenuState for transparency.  This is where the
GetHiddenState method is used.  Now the main game continues to
render--but not play--while the game is paused.  Note that this has
not been thoroughly tested yet, and may still cause bugs.