updated for Windows building
Created a winmake.sh script, which will automatically build for
Windows and copy over all of the .dll files and necessary
directories.  May add a function to automatically compress the
file into a .zip.
The win directory is also now part of the .gitignore file, so it
will not accidentally be committed.