Updated stack handling, bugfixed winmake
Stack handling was, in most cases, calling the game object directly
when it should have been calling the State's method, which calls
the game object.  This has been fixed.
Added the PopStates(int i) method, which will pop a number of
states at a time, fixing problems with jumping from the
PauseMenuState to the MainMenuState, popping the MainGameState
in between.
winmake was copying the openal dll as .win, rather than copying it
to the ./win directory.  I know, I'm an idiot.