updated winmake, fixed falling, added rounding, updated scaleFactor
winmake was updated, now copies over necessary files
Falling was implemented, and now a player is falling when he has
no floors under him.
Placing a shape or core now snaps to the nearest meter.
scaleFactor can be increased or decreased with the mouse wheel,
and has a catch to prevent a negative scale.