Added scaleFactor reset, updated saving
Added the ability to reset the scaleFactor to default upon pressing
the middle mouse button.
Updated the hotkeys for saving and loading stages to Ctrl-S and
Ctrl-O (O as in "Open"), respectively.
Fixed a bug which caused the game to crash if you tried to load a
file that didn't exist (because it was still deleting the present
stage first).