added exit/death conditions, updated saving to reflect this
The FixtureType enum now begins with 1 instead of 0, to prevent
the user data from returning a false bool value when evaluated.
There is now an exit object, of the Shape class, and coming into
contact with it for now simply kicks the player out of the stage;
it will eventually load the next stage.
The player will now die upon contact with any core, which simply
reloads test.bin for now; will reload whichever the current stage
is when there is a standard set of stages.
Saving now also saves and loads the exit object, and is therefore
now incompatible with previous releases.