Changed floor handling.
Floors are now handled with a vector of floors, rather than an
8-bit integer.  The vector contains a pointer to all of the floor
objects (which must all be Shape pointers for now).  This will be
used, along with b2Distance, to maintain player orientation when
the player is grounded.  This can be very useful, as it opens
possibilities for non-square shapes as floor objects.